requirements.txt
python-dotenv
pymysql
pypdf
openai>=1.30.0
